requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-jazzmin==2.6.0
drf-yasg==1.21.7
djoser==2.2.2
drf-writable-nested==0.7.0
django-filter==23.5
django-cors-headers==4.3.1
Pillow==10.1.0
channels==4.0.0
daphne==4.1.0
django-model-utils==4.4.0
python-dotenv==1.0.1
cloudinary==1.39.1